move tasks to main project.
