+++ Src/MoneyFox.Application/CloudBackup/BackupService.cs
@@ -70,7 +70,7 @@ namespace MoneyFox.Application.CloudBackup
 
         private readonly CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
         private readonly SemaphoreSlim semaphoreSlim = new SemaphoreSlim(1, 1);
-        private readonly NLog.Logger logManager = LogManager.GetCurrentClassLogger();
+        private readonly Logger logManager = LogManager.GetCurrentClassLogger();
 
         public BackupService(ICloudBackupService cloudBackupService,
                              IFileStore fileStore,
