+++ Src/MoneyFox.Ios/MoneyFox.iOS.csproj
@@ -245,7 +245,6 @@
     <None Include="Entitlements.plist" />
     <None Include="Info.plist" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Src\BackgroundTaskManager.cs" />
     <None Include="MoneyFox.iOS.ruleset" />
     <BundleResource Include="Resources\materialdesignicons.ttf">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
