+++ Src/MoneyFox.Uwp.Tasks/MoneyFox.Uwp.Tasks.csproj
@@ -132,9 +132,6 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ClearPaymentsTask.cs" />
-    <Compile Include="RecurringPaymentTask.cs" />
-    <Compile Include="SyncBackupTask.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Identity.Client">
