+++ Src/MoneyFox.Uwp/BackgroundTasks/ClearPaymentsTask.cs
@@ -7,18 +7,21 @@ using MoneyFox.Application.Facades;
 using MoneyFox.BusinessDbAccess.PaymentActions;
 using MoneyFox.BusinessLogic.PaymentActions;
 using MoneyFox.Persistence;
+using NLog;
 
-namespace MoneyFox.Uwp.Tasks
+namespace MoneyFox.Uwp.BackgroundTasks
 {
     /// <summary>
     ///     Background task to periodically clear payments.
     /// </summary>
     public sealed class ClearPaymentsTask : IBackgroundTask
     {
+        private readonly Logger logManager = LogManager.GetCurrentClassLogger();
+
         public async void Run(IBackgroundTaskInstance taskInstance)
         {
             BackgroundTaskDeferral deferral = taskInstance.GetDeferral();
-            Debug.WriteLine("ClearPayment started");
+            logManager.Debug("ClearPayment started");
             ExecutingPlatform.Current = AppPlatform.UWP;
 
             var settingsFacade = new SettingsFacade(new SettingsAdapter());
@@ -32,13 +35,12 @@ namespace MoneyFox.Uwp.Tasks
             }
             catch (Exception ex)
             {
-                Debug.Write(ex);
-                Debug.WriteLine("ClearPaymentTask stopped due to an error.");
+                logManager.Warn(ex, "ClearPaymentTask stopped due to an error.");
             }
             finally
             {
                 settingsFacade.LastExecutionTimeStampClearPayments = DateTime.Now;
-                Debug.WriteLine("ClearPaymentTask finished.");
+                logManager.Debug("ClearPaymentTask finished.");
                 deferral.Complete();
             }
         }
