+++ Src/MoneyFox.Uwp/BackgroundTasks/RecurringPaymentTask.cs
@@ -4,21 +4,24 @@ using MoneyFox.Application.Facades;
 using MoneyFox.BusinessDbAccess.PaymentActions;
 using MoneyFox.BusinessLogic.PaymentActions;
 using MoneyFox.Persistence;
+using NLog;
 using System;
 using System.Diagnostics;
 using Windows.ApplicationModel.Background;
 
-namespace MoneyFox.Uwp.Tasks
+namespace MoneyFox.Uwp.BackgroundTasks
 {
     /// <summary>
     /// Periodically tests if there are new recurring payments and creates these.
     /// </summary>
     public sealed class RecurringPaymentTask : IBackgroundTask
     {
+        private readonly Logger logManager = LogManager.GetCurrentClassLogger();
+
         public async void Run(IBackgroundTaskInstance taskInstance)
         {
             BackgroundTaskDeferral deferral = taskInstance.GetDeferral();
-            Debug.WriteLine("RecurringPaymentTask started.");
+            logManager.Debug("RecurringPaymentTask started.");
 
             ExecutingPlatform.Current = AppPlatform.UWP;
             var settingsFacade = new SettingsFacade(new SettingsAdapter());
@@ -32,13 +35,12 @@ namespace MoneyFox.Uwp.Tasks
             }
             catch(Exception ex)
             {
-                Debug.Write(ex);
-                Debug.WriteLine("RecurringPaymentTask stopped due to an error.");
+                logManager.Warn(ex, "RecurringPaymentTask stopped due to an error.");
             }
             finally
             {
                 settingsFacade.LastExecutionTimeStampRecurringPayments = DateTime.Now;
-                Debug.WriteLine("RecurringPaymentTask finished.");
+                logManager.Debug("RecurringPaymentTask finished.");
                 deferral.Complete();
             }
         }
